+++ src/main/java/codesquad/user/MemoryUserDao.java
@@ -1,4 +1,4 @@
-package codesquad.dao.user;
+package codesquad.user;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,11 +8,12 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicLong;
 
-public class UserDao {
+public class MemoryUserDao implements UserDao {
     private static final ConcurrentMap<Long, User> database = new ConcurrentHashMap<>();
-    private static final Logger log = LoggerFactory.getLogger(UserDao.class);
+    private static final Logger log = LoggerFactory.getLogger(MemoryUserDao.class);
     private static AtomicLong sequence = new AtomicLong(0);
 
+    @Override
     public User save(User user) {
         if (user.getId() == null) {
             user.setId(sequence.incrementAndGet());
@@ -21,10 +22,12 @@ public class UserDao {
         return user;
     }
 
+    @Override
     public User findById(Long id) {
         return database.get(id);
     }
 
+    @Override
     public User findByUserId(String userId) {
         log.debug("database: {}", database.values());
         return database.values().stream()
@@ -33,6 +36,7 @@ public class UserDao {
                 .orElse(null);
     }
 
+    @Override
     public List<User> findAll() {
         return List.copyOf(database.values());
     }
